[android] Sign development APK with Signature Scheme v2
Google now requires APK's to be to be signed with APK Signature Scheme v2(SHA256).
https://developer.android.com/about/versions/nougat/android-7.0.html#apk_signature_v2

If you sign your app using APK Signature Scheme v2 and make further
changes to the app, the app's signature is invalidated. For this reason,
zipalign should be called before signing the APK Signature Scheme v2,
not after.